--- app/auth/auth.css
@@ -1,127 +1,0 @@
-body {
-  box-shadow: inset 0 0 12px #1e83e7, inset 0 0 8px #41ea9a;
-}
-
-.main-content {
-  @apply absolute z-10 flex flex-col items-center justify-center w-full h-full;
-}
-
-.title-sign {
-  background: -webkit-linear-gradient(#1ce486, #1e83e7, #10c973);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  @apply py-2 pl-8 mb-8 font-mono text-xl font-black tracking-wide md:text-3xl;
-}
-
-.title-login {
-  background: -webkit-linear-gradient(#41ea9a, #1e83e7);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  @apply flex items-center pb-6 text-4xl font-black tracking-widest;
-}
-
-.btn-login-redirect {
-  @apply text-white bg-[#3290EF] hover:bg-[#4b9cec] focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 focus:outline-none;
-  @apply text-xl tracking-wider;
-}
-
-.or {
-  @apply my-2 text-3xl font-extrabold tracking-wider max-md:text-2xl;
-}
-
-.rights {
-  @apply absolute bottom-0 w-full mb-1.5 text-sm text-center md:text-xl;
-}
-
-
-.waves-top {
-  @apply top-0 rotate-180;
-
-}
-
-.waves-bottom {
-  @apply bottom-0;
-}
-
-.waves {
-  @apply absolute w-full min-h-[106px] max-h-[350px]
-}
-
-.parallax>use {
-  animation: waves 20s cubic-bezier(.65, .5, .45, .9) infinite;
-}
-
-.parallax>use:nth-child(1) {
-  animation-delay: -2s;
-  animation-duration: 7s;
-}
-
-.parallax>use:nth-child(2) {
-  animation-delay: -3s;
-  animation-duration: 10s;
-}
-
-.parallax>use:nth-child(3) {
-  animation-delay: -4s;
-  animation-duration: 13s;
-}
-
-.parallax>use:nth-child(4) {
-  animation-delay: -5s;
-  animation-duration: 18s;
-}
-
-@keyframes waves {
-  0% {
-    transform: translate3d(-90px, 0, 0);
-  }
-
-  100% {
-    transform: translate3d(95px, 0, 0);
-  }
-}
-
-
-
-.form-login-student {
-  @apply space-y-5;
-}
-
-.input-group {
-  @apply relative py-4
-}
-
-.input-field {
-  border: 0;
-  border-bottom: 2px solid #1e83e7;
-  @apply text-[#1e83e7] py-1.5 md:text-xl bg-transparent transition-[border-color] outline-0 placeholder:text-transparent;
-}
-
-.input-field:placeholder-shown ~ .input-label {
-  @apply text-xl top-5 cursor-text;
-}
-
-.input-label {
-  @apply absolute top-0 block text-base duration-300 text-primary;
-}
-
-.input-field:focus {
-  border-image: linear-gradient(to right, #41ea9a, #1e83e7, #9782e0);
-  border-image-slice: 1;
-  @apply border-4;
-}
-
-.input-field:focus ~ .input-label {
-  @apply absolute top-0 block duration-300 font-bold text-[#11998e] text-base ;
-}
-
-.input-error {
-  @apply absolute block py-px text-xs text-red-500 -bottom-1 whitespace-nowrap;
-}
-
-.btn-login {
-  background-image: linear-gradient(60deg,#41ea9a, #1e83e7);
-  @apply text-white hover:bg-[#4b9cec] mx-auto block  font-medium rounded-lg text-sm px-8 py-1.5;
-  @apply focus:ring-4 focus:ring-blue-300 focus:outline-none;
-  @apply text-xl tracking-widest;
-}
