+++ app/globals.css
@@ -1,6 +1,7 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+
 
 .loader {
   @apply fixed top-0 bottom-0 left-0 right-0 z-50 flex items-center justify-center text-center bg-black;
@@ -34,42 +35,4 @@
   50% {
     color: #000;
   }
-}
-
-
-
-
-
-.header {
-  @apply h-screen;
-}
-.negative {
-  @apply flex justify-between px-[3%] py-1.5 shadow shadow-stone-600/50 max-h-[6vh] max-md:relative;
-}
-
-.nav-links {
-  @apply flex transition items-center justify-center space-x-8 max-md:space-x-0 max-md:space-y-5
-    max-md:border-4 max-md:border-x-white max-md:border-y-blue-700 rounded-xl max-md:flex-col
-    max-md:items-stretch max-md:absolute max-md:top-[85%] max-md:left-[50%]
-    max-md:py-5 dark:text-white max-md:hidden max-md:bg-body-light max-md:dark:bg-body-dark z-50;
-}
-
-.nav-link {
-  @apply p-2 ease-in-out border-white max-md:dark:border-x-white hover:border-y-2 rounded-xl max-md:shadow max-md:rounded max-md:border-x-4 max-md:border-x-gray-800 max-md:inline-block max-md:w-full max-md:px-8 max-md:py-1;
-}
-
-.nav-link.active {
-  @apply font-bold text-blue-700 border-blue-600 lg:border-b;
-}
-
-.btn-auth {
-  @apply px-4 py-1.5 text-sm font-bold text-white transition bg-black dark:bg-gray-50 dark:text-black hover:bg-gray-800 dark:hover:bg-gray-300 rounded-xl flex items-center cursor-pointer;
-}
-
-.btn-search {
-  @apply p-2 text-center transition duration-150 ease-out hover:bg-gray-900 dark:text-white hover:text-white hover:dark:bg-slate-800 rounded-xl max-md:dark:border-x-white max-md:shadow max-md:rounded max-md:border-x-4 max-md:border-x-gray-900 max-md:p-1;
-}
-
-.btn-theme {
-  @apply p-2 transition duration-150 ease-out hover:bg-gray-900 hover:text-white dark:text-white hover:dark:bg-slate-800 rounded-xl max-md:dark:border-x-white max-md:shadow max-md:rounded max-md:border-x-4 max-md:border-x-gray-900 max-md:p-1;
 }
