+++ pages/auth/index.tsx
@@ -1,0 +1,16 @@
+import { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: 'Auth',
+  description: 'The Auth page'
+}
+
+export default function Auth() {
+  
+
+  return (
+    <>
+      <h1>Auth</h1>
+    </>
+  )
+}
