+++ pages/auth/login.tsx
@@ -1,4 +1,4 @@
-import type { Metadata } from "next"
+import { Metadata } from "next"
 
 export const metadata: Metadata = {
   title: 'Auth Login',
